examples/keras_with_grpc/requirements.txt
tensorflow
grpcio
grpcio-tools
